MF - MF prepare to create page and controllers working OK.
Made a nice Ajax integration that can be used(the same kind of integration) in other pages.
When the user selects the MFType in the form, it makes a $.get(ajax get) call to a controller giving the MFType selected as a GET parameter.
This controller then get what is the corresponding number of parameters for this MFType and return it JSON format.
Back in the cliente it receive this number of parameters and add to the form the corresponding number of params input.

Yeaaaah o/

OBS: to make this better, I changed the html5boilerplate to have the Jquery import in the header, this way any Jquery method can be used anywhare in the page.
And so all this js was added to the end of the prepareCreate.scala.html page of the MF. This way all the other pages wont get saturated with js that are not in use.

refs #15