OperatorSet - Now the prepareCreate view has a js script that for each Operator that the user select it does a ajax get in the url '''/ajax/get_num_params_operator_option''' passing the the id_option and the id_opType. The server returns a list of strings(that represent the name of each parameter this selected option of this operator has. If no parameters or if the operator is the default one, it returns a empty list.
Now I need to do the js part that will create the new forms input for the parameters using the return of the ajaxGet paramsList.
Probably it's a good idea to create a panel(that wont allow the user to do anything in the rest of the site until closed) with this input appearing.
When the user press 'ok' it will make the inputs hidden in the form.
refs #20 #25